extract run artifact route
